Added orientation support
iPads can alter orientation to be upside down so I added support. It’s more of a comfort feature, and I have no idea if it works because I’m using the iOS 13 framework because I only have an IPad on that version so the library’s might be different  on it. I’m planning on getting another IPad or a corellium subscription so i will test it then.